+++ tests/Datasets/UnlabeledTest.php
@@ -520,6 +520,7 @@ class UnlabeledTest extends TestCase
                 'offset' => 3,
                 'type' => 'continuous',
                 'mean' => 0.3333333333333333,
+                'variance' => 9.792222222222222,
                 'standard deviation' => 3.129252661934191,
                 'skewness' => -0.4481030843690633,
                 'kurtosis' => -1.1330702741786107,